mycloud/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9
python-dotenv>=1.0
cryptography>=42.0
django-cors-headers>=4.3
